Clean up parsing/allocator leaks in svelte_analyze
- Rename parse_js.rs → analyze_semantic.rs (module does scope registration, not parsing)
- ComponentScoping::empty(): use Scoping::default() instead of parsing empty string
- Extract analyze_module parsing into svelte_parser::parse_module()
- Move enrich_from_unresolved to svelte_parser::script_info
- Move oxc_parser and oxc_allocator to dev-dependencies in svelte_analyze
