+++ crates/svelte_analyze/src/analyze_semantic.rs
