+++ crates/svelte_parser/Cargo.toml
@@ -10,6 +10,7 @@ svelte_diagnostics = { path = "../svelte_diagnostics" }
 oxc_allocator = { workspace = true }
 oxc_ast = { workspace = true }
 oxc_parser = { workspace = true }
+oxc_semantic = { workspace = true }
 oxc_span = { workspace = true }
 oxc_syntax = { workspace = true }
 compact_str = { workspace = true }
